+++ src/features/usage-consolidation-accordion/summary-table/summary-table.js
@@ -4,7 +4,6 @@ import {
 } from 'react';
 import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
-import noop from 'lodash/noop';
 
 import {
   MultiColumnList,
@@ -24,11 +23,8 @@ const propTypes = {
   customProperties: PropTypes.object,
   entityType: PropTypes.string.isRequired,
   id: PropTypes.string.isRequired,
-  isExportDisabled: PropTypes.bool,
   metricType: PropTypes.string,
   noCostPerUseAvailable: PropTypes.bool.isRequired,
-  onExportTitles: PropTypes.func,
-  onViewTitles: PropTypes.func,
 };
 
 const SummaryTable = ({
@@ -36,11 +32,8 @@ const SummaryTable = ({
   customProperties,
   id,
   costPerUseType,
-  onViewTitles,
   entityType,
   metricType,
-  onExportTitles,
-  isExportDisabled,
   noCostPerUseAvailable,
   ...rest
 }) => {
@@ -87,12 +80,6 @@ const SummaryTable = ({
       />
     </KeyValue>
   );
-};
-
-SummaryTable.defaultProps = {
-  isExportDisabled: false,
-  onExportTitles: noop,
-  onViewTitles: noop,
 };
 
 SummaryTable.propTypes = propTypes;